requirements.txt
requests~=2.32.5
python-dotenv~=1.2.1
anthropic~=0.73.0
markdown~=3.10
pyyaml~=6.0.3
defeatbeta-api~=0.0.24
pytest~=9.0.1
pytest-cov~=7.0.0
ruamel.yaml~=0.18.16
pre-commit~=4.4.0
black~=25.11.0
isort~=7.0.0
flake8~=7.3.0
